AI updates
Fixed an issue that caused not moving units in doorways first as intended.
The assignment of scouting-duties now takes unit-proximity into account. This shall lead to more effective and more efficient scouting.
Removed cases where units didn't scout at all in order to prevent showing up on movement-scanners. So AI will camp less and more proactively try to find player units on their turn, even if that helps the player to find them.
Reverted the hiding logic to how it used to work in the past. This means AI should be better at avoiding to be caught out in the open and generally a bit less aggressive and more sneaky.
Fixed an bug that caused the AI not to open doors when assuming there's an enemy on the other side. Note: This makes fighting near doors, such as inside UFOs significantly more dangerous.
The damage-esitmation-logic of the AI now is aware of the concept of overkill and whether a weapons' damage per time-unit comes mostly from multiple attacks against the same target or allows attacking different targets with multiple attacks.
This leads to a significant reduction of the behavior of getting as close as possible before attacking and helps the AI to avoid overkill on a single unit and getting stranded with too little timeunits to hide again.